feat: Port NuGetResolver.UI to AvaloniaUI
