+++ sources/shared/Stride.NuGetResolver.UI/SplashScreenWindow.axaml
@@ -1,18 +1,17 @@
 <Window
-    x:Class="Stride.NuGetResolver.SplashScreenWindow"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns="https://github.com/avaloniaui"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:Stride.NuGetResolver"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
+    x:Class="Stride.NuGetResolver.SplashScreenWindow"
     WindowStartupLocation="CenterScreen"
-    WindowStyle="None"
-    AllowsTransparency="True"
+	  SystemDecorations="None"
     Background="#1A1A1D"
     Width="480"
     Height="380"
-    ResizeMode="NoResize">
+    CanResize="False">
 
   <Window.Resources>
     <ResourceDictionary>
@@ -25,11 +24,8 @@
       </DrawingImage>
     </ResourceDictionary>
   </Window.Resources>
-  <Window.CommandBindings>
-    <CommandBinding Command="ApplicationCommands.Close" Executed="CloseCommand"/>
-  </Window.CommandBindings>
   <Grid>
-    <Button HorizontalAlignment="Right" VerticalAlignment="Top" Foreground="#FFFFFF" Background="#1A1A1D" Command="{x:Static ApplicationCommands.Close}" Content="X" Padding="4" x:Name="CloseButton" Visibility="Hidden" />
+    <Button HorizontalAlignment="Right" VerticalAlignment="Top" Foreground="#FFFFFF" Background="#1A1A1D" Click="CloseCommand" Content="X" Padding="4" x:Name="CloseButton" IsVisible="False" />
     <StackPanel Orientation="Vertical" Margin="10">
       <Image x:Name="Logo"
            HorizontalAlignment="Center" VerticalAlignment="Center"
@@ -37,7 +33,7 @@
            Margin="10"/>
       <ProgressBar IsIndeterminate="True" Width="424" Margin="10" />
       <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center" Foreground="#FFFFFF" Margin="10" x:Name="Message">Restoring NuGet packages...</TextBlock>
-      <TextBox x:Name="Log" Width="424" Height="120" HorizontalAlignment="Center" VerticalAlignment="Center" Foreground="#C0C0C0" Background="#1A1A1D" Margin="10" AcceptsReturn="True" VerticalScrollBarVisibility="Auto" />
+      <TextBox x:Name="Log" Width="424" Height="120" IsReadOnly="True" HorizontalAlignment="Center" VerticalAlignment="Center" Foreground="#C0C0C0" Background="#1A1A1D" Margin="10" AcceptsReturn="True" />
     </StackPanel>
   </Grid>
 </Window>
